Removed Link routing from Button and extended Button flexibility.
