+++ client/src/components/Buttons/index.js
@@ -1,1 +1,1 @@
-export { default as LinkButton } from './LinkButton';
+export { default as RoundButton } from './RoundButton';
